figured out multiple inheritance
just needed the file_name specified explicitly. It wasn't a python thing, it was a me thing. You can also call the init on both classes manually and that works, and may actually be more clear which I may do, but this works fine I think. It probably could be slightly more efficient to do it manually but I don't have enough images for it to make a difference